+++ modules/apps/journal/journal-web/src/main/resources/META-INF/resources/init.jsp
@@ -54,8 +54,6 @@ page import="com.liferay.dynamic.data.mapping.service.DDMStructureLocalServiceUt
 page import="com.liferay.dynamic.data.mapping.service.DDMTemplateLocalServiceUtil" %><%@
 page import="com.liferay.dynamic.data.mapping.service.DDMTemplateServiceUtil" %><%@
 page import="com.liferay.dynamic.data.mapping.util.DDMNavigationHelper" %><%@
-page import="com.liferay.dynamic.data.mapping.validator.DDMFormLayoutValidationException" %><%@
-page import="com.liferay.dynamic.data.mapping.validator.DDMFormValidationException" %><%@
 page import="com.liferay.exportimport.kernel.exception.ExportImportContentValidationException" %><%@
 page import="com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem" %><%@
 page import="com.liferay.frontend.taglib.clay.servlet.taglib.util.JSPDropdownItemList" %><%@
