+++ modules/apps/journal/journal-web/src/main/resources/META-INF/resources/js/ChangeDefaultLanguage.es.js
@@ -16,7 +16,7 @@ class ChangeDefaultLanguage extends Component {
 
 		this.defaultLanguage = defaultLanguage;
 		this.languages = this.languages.map(
-			language => {
+			language => {
 				return Object.assign(
 					{},
 					language,
